requirements.txt
python-dotenv
openai
streamlit-chat